desktop-app/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
视频工具平台 - 桌面版应用
主入口文件
"""

import sys
import os
from PyQt6.QtWidgets import QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QLabel, QTabWidget, QTextEdit, QFileDialog, QMessageBox
from PyQt6.QtCore import Qt, QThread, pyqtSignal
from PyQt6.QtGui import QFont, QIcon, QPixmap
import cv2
from PIL import Image
import numpy as np

class VideoToolsApp(QMainWindow):
    def __init__(self):
        super().__init__()
        self.setWindowTitle("视频工具平台 v1.0.0")
        self.setGeometry(100, 100, 1200, 800)
        
        # 设置应用图标
        self.setWindowIcon(QIcon("icon.ico"))
        
        # 创建主界面
        self.init_ui()
        
    def init_ui(self):
        """初始化用户界面"""
        central_widget = QWidget()
        self.setCentralWidget(central_widget)
        
        # 主布局
        main_layout = QVBoxLayout(central_widget)
        
        # 标题
        title_label = QLabel("视频工具平台")
        title_label.setFont(QFont("Arial", 24, QFont.Weight.Bold))
        title_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
        title_label.setStyleSheet("color: #2563eb; margin: 20px;")
        main_layout.addWidget(title_label)
        
        # 功能标签页
        self.tab_widget = QTabWidget()
        main_layout.addWidget(self.tab_widget)
        
        # 创建各个功能页面
        self.create_image_tools_tab()
        self.create_screenshot_tab()
        self.create_video_tools_tab()
        
    def create_image_tools_tab(self):
        """创建图片处理工具页面"""
        tab = QWidget()
        layout = QVBoxLayout(tab)
        
        # 标题
        title = QLabel("🖼️ 图片处理工具")
        title.setFont(QFont("Arial", 16, QFont.Weight.Bold))
        layout.addWidget(title)
        
        # 按钮区域
        button_layout = QHBoxLayout()
        
        # 图片缩放按钮
        resize_btn = QPushButton("图片缩放")
        resize_btn.clicked.connect(self.resize_image)
        button_layout.addWidget(resize_btn)
        
        # 格式转换按钮
        convert_btn = QPushButton("格式转换")
        convert_btn.clicked.connect(self.convert_image_format)
        button_layout.addWidget(convert_btn)
        
        # 图片压缩按钮
        compress_btn = QPushButton("图片压缩")
        compress_btn.clicked.connect(self.compress_image)
        button_layout.addWidget(compress_btn)
        
        layout.addLayout(button_layout)
        
        # 日志区域
        self.image_log = QTextEdit()
        self.image_log.setMaximumHeight(200)
        layout.addWidget(self.image_log)
        
        self.tab_widget.addTab(tab, "图片处理")
        
    def create_screenshot_tab(self):
        """创建截屏工具页面"""
        tab = QWidget()
        layout = QVBoxLayout(tab)
        
        # 标题
        title = QLabel("📸 截屏工具")
        title.setFont(QFont("Arial", 16, QFont.Weight.Bold))
        layout.addWidget(title)
        
        # 按钮区域
        button_layout = QHBoxLayout()
        
        # 全屏截图按钮
        fullscreen_btn = QPushButton("全屏截图")
        fullscreen_btn.clicked.connect(self.fullscreen_screenshot)
        button_layout.addWidget(fullscreen_btn)
        
        # 区域截图按钮
        area_btn = QPushButton("区域截图")
        area_btn.clicked.connect(self.area_screenshot)
        button_layout.addWidget(area_btn)
        
        # 窗口截图按钮
        window_btn = QPushButton("窗口截图")
        window_btn.clicked.connect(self.window_screenshot)
        button_layout.addWidget(window_btn)
        
        layout.addLayout(button_layout)
        
        # 日志区域
        self.screenshot_log = QTextEdit()
        self.screenshot_log.setMaximumHeight(200)
        layout.addWidget(self.screenshot_log)
        
        self.tab_widget.addTab(tab, "截屏工具")
        
    def create_video_tools_tab(self):
        """创建视频处理工具页面"""
        tab = QWidget()
        layout = QVBoxLayout(tab)
        
        # 标题
        title = QLabel("🎬 视频处理工具")
        title.setFont(QFont("Arial", 16, QFont.Weight.Bold))
        layout.addWidget(title)
        
        # 按钮区域
        button_layout = QHBoxLayout()
        
        # 视频格式转换按钮
        video_convert_btn = QPushButton("视频格式转换")
        video_convert_btn.clicked.connect(self.convert_video_format)
        button_layout.addWidget(video_convert_btn)
        
        # 音频提取按钮
        audio_extract_btn = QPushButton("音频提取")
        audio_extract_btn.clicked.connect(self.extract_audio)
        button_layout.addWidget(audio_extract_btn)
        
        # 视频压缩按钮
        video_compress_btn = QPushButton("视频压缩")
        video_compress_btn.clicked.connect(self.compress_video)
        button_layout.addWidget(video_compress_btn)
        
        layout.addLayout(button_layout)
        
        # 日志区域
        self.video_log = QTextEdit()
        self.video_log.setMaximumHeight(200)
        layout.addWidget(self.video_log)
        
        self.tab_widget.addTab(tab, "视频处理")
    
    # 图片处理功能
    def resize_image(self):
        """图片缩放功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择图片", "", "图片文件 (*.jpg *.jpeg *.png *.bmp)")
        if file_path:
            try:
                # 这里实现图片缩放逻辑
                self.image_log.append(f"正在处理图片: {file_path}")
                # TODO: 实现具体的缩放功能
                QMessageBox.information(self, "成功", "图片缩放功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"处理失败: {str(e)}")
    
    def convert_image_format(self):
        """图片格式转换功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择图片", "", "图片文件 (*.jpg *.jpeg *.png *.bmp)")
        if file_path:
            try:
                self.image_log.append(f"正在转换图片格式: {file_path}")
                # TODO: 实现具体的格式转换功能
                QMessageBox.information(self, "成功", "格式转换功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"转换失败: {str(e)}")
    
    def compress_image(self):
        """图片压缩功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择图片", "", "图片文件 (*.jpg *.jpeg *.png *.bmp)")
        if file_path:
            try:
                self.image_log.append(f"正在压缩图片: {file_path}")
                # TODO: 实现具体的压缩功能
                QMessageBox.information(self, "成功", "图片压缩功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"压缩失败: {str(e)}")
    
    # 截屏功能
    def fullscreen_screenshot(self):
        """全屏截图功能"""
        try:
            self.screenshot_log.append("正在执行全屏截图...")
            # TODO: 实现全屏截图功能
            QMessageBox.information(self, "成功", "全屏截图功能开发中...")
        except Exception as e:
            QMessageBox.critical(self, "错误", f"截图失败: {str(e)}")
    
    def area_screenshot(self):
        """区域截图功能"""
        try:
            self.screenshot_log.append("正在执行区域截图...")
            # TODO: 实现区域截图功能
            QMessageBox.information(self, "成功", "区域截图功能开发中...")
        except Exception as e:
            QMessageBox.critical(self, "错误", f"截图失败: {str(e)}")
    
    def window_screenshot(self):
        """窗口截图功能"""
        try:
            self.screenshot_log.append("正在执行窗口截图...")
            # TODO: 实现窗口截图功能
            QMessageBox.information(self, "成功", "窗口截图功能开发中...")
        except Exception as e:
            QMessageBox.critical(self, "错误", f"截图失败: {str(e)}")
    
    # 视频处理功能
    def convert_video_format(self):
        """视频格式转换功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择视频", "", "视频文件 (*.mp4 *.avi *.mov *.mkv)")
        if file_path:
            try:
                self.video_log.append(f"正在转换视频格式: {file_path}")
                # TODO: 实现具体的视频格式转换功能
                QMessageBox.information(self, "成功", "视频格式转换功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"转换失败: {str(e)}")
    
    def extract_audio(self):
        """音频提取功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择视频", "", "视频文件 (*.mp4 *.avi *.mov *.mkv)")
        if file_path:
            try:
                self.video_log.append(f"正在提取音频: {file_path}")
                # TODO: 实现具体的音频提取功能
                QMessageBox.information(self, "成功", "音频提取功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"提取失败: {str(e)}")
    
    def compress_video(self):
        """视频压缩功能"""
        file_path, _ = QFileDialog.getOpenFileName(self, "选择视频", "", "视频文件 (*.mp4 *.avi *.mov *.mkv)")
        if file_path:
            try:
                self.video_log.append(f"正在压缩视频: {file_path}")
                # TODO: 实现具体的视频压缩功能
                QMessageBox.information(self, "成功", "视频压缩功能开发中...")
            except Exception as e:
                QMessageBox.critical(self, "错误", f"压缩失败: {str(e)}")

def main():
    app = QApplication(sys.argv)
    
    # 设置应用样式
    app.setStyle('Fusion')
    
    # 创建主窗口
    window = VideoToolsApp()
    window.show()
    
    sys.exit(app.exec())

if __name__ == "__main__":
    main()
