desktop-app/requirements.txt
PyQt6==6.6.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pyinstaller==6.2.0
